pyproject.toml
[tool.poetry]
name = "restaurantbookinglab-backend"
version = "0.1.0"
description = ""
authors = ["Sergio Vargas <svargas219@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
mysqlclient = "^2.2.0"
celery = "^5.3.4"
redis = "^5.0.1"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
